requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
braceexpand==0.1.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coloredlogs==15.0.1
discord.py==2.0.1
frozenlist==1.3.1
humanfriendly==10.0
humanize==4.3.0
idna==3.3
import-expression==1.1.4
jishaku==2.5.1
line-profiler==3.5.1
multidict==6.0.2
mystbin-py==5.0.2
Pillow==9.2.0
pycparser==2.21
PyNaCl==1.5.0
pyreadline3==3.4.1
six==1.16.0
toml==0.10.2
typing_extensions==4.3.0
wavelink==1.3.2
yarl==1.8.1
